added ui and connected the open weather to UI

+++ app/src/main/java/com/example/weatherapp/models/Sys.kt
@@ -3,7 +3,7 @@ package com.example.weatherapp.models
 data class Sys(
     val country: String,
     val id: Int,
-    val sunrise: Int,
-    val sunset: Int,
+    val sunrise: Long,
+    val sunset: Long,
     val type: Int
 )
